scripts/requirements.txt
python-docx
pandas
regex
